Moved package_output() to util.package_output() where it can be mocked
